refactor: split authorize benchmark runner
